A little refactoring to introduce a FollowersView so were are not duplicating the StreamsView
